+++ tests/test_Guards.py
@@ -329,6 +329,18 @@ def test_Guards__safer_getattr__5():
     ) == str(err.value)
 
 
+def test_Guards__safer_getattr_raise():
+    from types import SimpleNamespace
+
+    from RestrictedPython.Guards import safer_getattr_raise as ga
+
+    o = SimpleNamespace(a="a")
+    assert ga(o, "a") == "a"
+    assert ga(o, "b", None) is None
+    with pytest.raises(AttributeError):
+        ga(o, "b")
+
+
 def test_call_py3_builtins():
     """It should not be allowed to access global builtins in Python3."""
     result = compile_restricted_exec('builtins["getattr"]')